separate interface definition
