+++ src/store/card/types.ts
@@ -1,0 +1,10 @@
+import { Label } from '../label/types';
+
+
+export interface Card {
+  id: number
+  title: string
+  description: string
+  listId: number
+  labels: Label[]
+}
