+++ src/store/label/types.ts
@@ -1,0 +1,5 @@
+export interface Label {
+  id: number
+  name: string
+  color: string
+}
