+++ src/store/list/types.ts
@@ -1,0 +1,8 @@
+import { Card } from '../card/types';
+
+export interface List {
+  id: number
+  name: string
+  boardId: number
+  cards: Card[]
+}
